api/requirements.txt
flask==3.0.2
flask-cors==4.0.0
anthropic==0.18.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.5.2
gunicorn==21.2.0
requests==2.31.0
docx2txt==0.8
Werkzeug==3.0.1 